% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.1_Funcion_DENUE_Buffer.R
\name{denue_buffer}
\alias{denue_buffer}
\title{Consultas DENUE Buffer}
\usage{
denue_buffer(
  latitud,
  longitud,
  metros = 250,
  exportar = TRUE,
  keyword = "todos",
  token = "a3c33cba-ab0a-4439-8c01-a6cec4453f8f"
)
}
\arguments{
\item{latitud}{latitud.}

\item{longitud}{longitud.}

\item{metros}{metros de radio. Minimo 250 - Maximo 5000}

\item{exportar}{TRUE para exportar la informacion a xlsx.}

\item{keyword}{Palabra clave. Ejemplo "banca-multiple"}
}
\description{
Función para realizar consultas en un radio de hasta 5000 metros.
}
\details{
Esta funcion es parte del proceso de estudio de mercado. Permite
extraer todas las unidades economicas dentro de un radio de hasta 5000 metros.
}
\examples{
# Todos los cajeros automaticos en un radio de 2000 metros en Leon
\dontrun{
latitud <- 21.059960
longitud <- -101.574593
metros <- 2000
exportar <- FALSE
keyword <- "cajero-automatico"
atm <- denue_buffer(latitud, longitud, metros, exportar, keyword)
}

}
