% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.1_Funcion_DENUE_Localidad.R
\name{denue_localidad}
\alias{denue_localidad}
\title{DENUE Localidad}
\usage{
denue_localidad(
  cve_ent,
  localidad = "todos",
  exportar = TRUE,
  token = "a3c33cba-ab0a-4439-8c01-a6cec4453f8f"
)
}
\arguments{
\item{cve_ent}{Clave de la entidad a consultar.}

\item{localidad}{Nombre de la localidad a consultar.}

\item{exportar}{TRUE para exportar la informacion a xlsx.}
}
\description{
Esta funcion permite extraer todas las unidades economicas de una localidad.
}
\details{
Esta funcion es parte del proceso de estudio de mercado. Permite extraer
todas las unidades economicas de una localidad y segmentarlas para su posterior
analisis
}
\examples{
# Todas las unidades economicas de San Miguel de Allende, Guanajuato
\dontrun{
cve_ent <- 11
localidad <- "san-miguel-de-allende"
exportar <- FALSE
ue <- denue_localidad(cve_ent, localidad, exportar)
}

}
