% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06.0_Funcion_DENUE_Stats.R
\name{denue_stats}
\alias{denue_stats}
\title{DENUE Estadísticas}
\usage{
denue_stats(
  latitud_vector,
  longitud_vector,
  metros = 250,
  keyword = "todos",
  exportar = FALSE,
  token
)
}
\arguments{
\item{latitud_vector}{Vector de coordenada latitud.}

\item{metros}{Metros del radio.}

\item{keyword}{Palabra clave.}

\item{exportar}{TRUE para exportar la informacion a xlsx.}

\item{localidad_vector}{Vector de coordenada longitud.}
}
\description{
Genera estadisticas de las unidades economicas en un radio de 250 a 5000 metros
}
\details{
Esta funcion es parte del proceso de estudio de mercado. Permite extraer
estadisticas de las unidades economicas a partir de un par de coordenadas, en un radio
minimo de 250 metros y maximo de 5000 metros.
}
\examples{
# Estadisticas de unidades economicas en el centro de Oaxaca y Monterrey
\dontrun{
latitud_vector <- c(17.061857, 25.677952)
longitud_vector <- c(-96.724723, -100.313294)
metros <- 1500
exportar <- FALSE
ue <- denue_localidad(latitud_vector, longitud_vector, metros, exportar)
}

}
