% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.0_Read_Multiple_Excel_Sheets.R
\name{read_multiple_excel}
\alias{read_multiple_excel}
\title{Leer Multiples Hojas de Excel.}
\usage{
read_multiple_excel(path)
}
\arguments{
\item{path}{Ruta donde se encuentran los archivos.}
}
\description{
Leer todas las hojas de varios documentos de excel.
}
\details{
Esta funcion permite leer todas las hojas de varios documentos de excel.
}
\examples{
# Leer todas las hojas de los archivos en una carpeta
\dontrun{
data_df <- dir_ls(regexp = "xlsx") \%>\%
 map_df(read_multiple_excel,
        .id = "city")
}

}
